Issue 39296: x86-64 parameter zero/sign extension
In 286ad9d77, zero/sign extension of integer parameters on the x86-64 SYSV ABI
was disabled for 8 bit parameters. While compatible with the ABI, the ABI is
not explicit about whether or not this is required and not all compilers follow
this convention (or have not always done so).

For maximum compatibility with other compilers, FPC will now in case of 8/16
bit parameters/functionresults,
a) zero/sign-extend parameters both other caller and callee side
b) zero/sign extend the function result on the callee side only on Darwin
(becaus that platform needs to keep compatibility with an older version of
clang/llvm that did so